weather app/requirements.txt
flask
requests
twilio
gunicorn
python-dotenv
flask-cors